Move json_logger to logger.py
